refactor: reorganize auth and hero components
