+++ app/(auth)/sign-up/page.tsx
@@ -1,7 +1,7 @@
-import AuthCard from "@/common/components/auth-card";
-import ButtonHandler from "@/common/components/auth-card/sign-up/button-handler";
-import HightlightBar from "@/common/components/auth-card/sign-up/highlight-bar";
-import RegistrationStep from "@/common/components/auth-card/sign-up/registration-step";
+import AuthCard from "@/common/components/auth/auth-card";
+import ButtonHandler from "@/common/components/auth/auth-sign-up/button-handler";
+import HightlightBar from "@/common/components/auth/auth-sign-up/highlight-bar";
+import RegistrationStep from "@/common/components/auth/auth-sign-up/registration-step";
 import { SignUpFormProvider } from "@/common/providers/form-provider";
 
 export default function Page() {
