+++ common/components/auth/auth-sign-up/highlight-bar.tsx
