+++ common/components/auth/auth-sign-up/registration-step.tsx
