+++ common/components/auth/auth-sign-up/type-selection-form.tsx
