+++ common/components/auth/auth-sign-up/user-type-card.tsx
