+++ common/components/home/hero/hero-content/index.tsx
